+++ apps/dotcom/client/public/tla/locales-compiled/en.json
@@ -126,6 +126,7 @@
   "ec611e9a08": "Account menu",
   "ed53713239": "i18n: Accented",
   "ed75712b0e": "Duplicate",
+  "f0e8bff6f9": "User settings",
   "f2569594b2": "We use these cookies to save your files and settings.",
   "f2a6c498fb": "Delete",
   "f43fb62093": "Back to tldraw.",
