+++ apps/dotcom/client/public/tla/locales/en.json
@@ -381,6 +381,9 @@
   "ed75712b0e": {
     "translation": "Duplicate"
   },
+  "f0e8bff6f9": {
+    "translation": "User settings"
+  },
   "f2569594b2": {
     "translation": "We use these cookies to save your files and settings."
   },
